pyproject.toml
[tool.poetry]
name = "d1_baseball_api"
version = "0.1.0"
description = ""
authors = ["jmurrah <jacobhmurrah@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pytest = "^8.4.1"
python-dotenv = "^1.1.1"
beautifulsoup4 = "^4.13.5"
requests = "^2.32.5"
supabase = "^2.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
